Funcionan las teclas del shader Mandel Fractal.

+++ PlantillaOpenGL_Proyecto_Shader01/PlantillaOpenGL01/spirofield.frag
@@ -11,7 +11,7 @@ vec4 HSVtoRGB( float h, float s, float v ){
    }*/
    
    h /= 60;         // sector 0 to 5
-   i = floor( h );
+   i = int(floor( h ));
    f = h - i;         // factorial part of h
    p = v * ( 1 - s );
    q = v * ( 1 - s * f );
